1.2.1 - GUI Exploit + Credits Spent returning null for players who havent spent any credits yet. (New players)
