--- config.yml
@@ -1,86 +1,0 @@
-#============================================================================================================================| 
-#    _____                      .__   __________           .___                     
-#   /     \   _________________ |  |__\______   \ ____   __| _/____   ____   _____  
-#  /  \ /  \ /  _ \_  __ \____ \|  |  \|       _// __ \ / __ |/ __ \_/ __ \ /     \ 
-# /    Y    (  <_> )  | \/  |_> >   Y  \    |   \  ___// /_/ \  ___/\  ___/|  Y Y  \
-# \____|__  /\____/|__|  |   __/|___|  /____|_  /\___  >____ |\___  >\___  >__|_|  /
-#         \/             |__|        \/       \/     \/     \/    \/     \/      \/ 
-#
-#============================================================================================================================|
-#
-# Plugin Version: 1.2.0
-# Spigot Link: https://www.spigotmc.org/resources/morphredeem-mcmmo-credits-1-14.67435/
-# Discord Link: https://discord.gg/6bnU9xn
-# Author's Discord: Morphie#6969
-# 
-# If you have issues or questions feel free to message me on spigot OR discord.
-# Please do keep in mind however that I will respond when I can! <3
-#
-#============================================================================================================================|
-
-Settings:
-  # The color of glass in the main /morphreedem menu.
-  BackgroundGlassColor: 11
-  DisabledSkills:
-    Enabled: false
-    # Item to replace disabled Skills with.
-    ReplaceGUIItem:
-      Enabled: false
-      ItemName: STRUCTURE_VOID
-    # Skills that players won't be able to apply credits to.
-    SkillsToDisable:
-    - Acrobatics
-  # Enable or disable the plugin credits item in the /morphredeem menu.
-  PluginCredits:
-    Enabled: true
-  # Change the Item types in the main /morphredeem menu.
-  GUI:
-    Acrobatics:
-      ItemName: LEATHER_BOOTS
-    Alchemy:
-      ItemName: POTION
-    Archery:
-      ItemName: BOW
-    Axes:
-      ItemName: DIAMOND_AXE
-    Excavation:
-      ItemName: DIAMOND_SHOVEL
-    Fishing:
-      ItemName: FISHING_ROD 
-    Herbalism:
-      ItemName: FARMLAND
-    Mining:
-      ItemName: DIAMOND_PICKAXE
-    Repair:
-      ItemName: ANVIL
-    Swords:
-      ItemName: DIAMOND_SWORD
-    Taming:
-      ItemName: BONE
-    Unarmed:
-      ItemName: STICK
-    Woodcutting:
-      ItemName: OAK_WOOD
-    mcMMOCredits:
-      ItemName: NETHER_STAR
-    CreditsSpent:
-      ItemName: EMERALD
-    PluginCredits:
-      ItemName: PAPER
-
-#============================================================================================================================|
-
-# Storage method, can be MySQL or YML
-StorageMethod: "YML"
-
-# Requires 'StorageMethod' to be MySQL
-MySQL:
-  Username: "root"
-  Password: "password"
-  Host: "localhost"
-  Port: 3306
-  SSL: false
-  Database: "minecraft"
-  TablePrefix: "mr_"
-  
-#============================================================================================================================|
