--- plugin.yml
@@ -1,12 +1,0 @@
-name: MorphRedeem
-version: 1.2.0
-main: net.naturva.morphie.mr.MorphRedeem
-author: Morphie
-api-version: 1.13
-depend: [mcMMO]
-softdepend: [PlaceholderAPI]
-commands:
-   morphredeem:
-    description: Main command for the MorphRedeem plugin!
-    usage: /<command>
-    aliases: [mr, redeem]
